// src/schema.ts
const factDbcTable = {
    "name": "factDbc",
    "columns": {
        "jaar": {
            "type": "number",
            "optional": false,
            "customType": null as unknown as number
        },
        "behandelendSpecialismeCd": {
            "type": "string",
            "optional": false,
            "customType": null as unknown as string,
            "serverName": "behandelend_specialisme_cd"
        },
        "typerendeDiagnoseCd": {
            "type": "string",
            "optional": false,
            "customType": null as unknown as string,
            "serverName": "typerende_diagnose_cd"
        },
        "zorgproductCd": {
            "type": "string",
            "optional": false,
            "customType": null as unknown as string,
            "serverName": "zorgproduct_cd"
        },
        "aantalPatPerZpd": {
            "type": "number",
            "optional": false,
            "customType": null as unknown as number,
            "serverName": "aantal_pat_per_zpd"
        },
        "aantalSubtrajectPerZpd": {
            "type": "number",
            "optional": false,
            "customType": null as unknown as number,
            "serverName": "aantal_subtraject_per_zpd"
        },
        "aantalPatPerDiag": {
            "type": "number",
            "optional": false,
            "customType": null as unknown as number,
            "serverName": "aantal_pat_per_diag"
        },
        "aantalSubtrajectPerDiag": {
            "type": "number",
            "optional": false,
            "customType": null as unknown as number,
            "serverName": "aantal_subtraject_per_diag"
        },
        "aantalPatPerSpc": {
            "type": "number",
            "optional": false,
            "customType": null as unknown as number,
            "serverName": "aantal_pat_per_spc"
        },
        "aantalSubtrajectPerSpc": {
            "type": "number",
            "optional": false,
            "customType": null as unknown as number,
            "serverName": "aantal_subtraject_per_spc"
        },
        "gemiddeldeVerkoopprijs": {
            "type": "number",
            "optional": true,
            "customType": null as unknown as number,
            "serverName": "gemiddelde_verkoopprijs"
        },
        "versie": {
            "type": "string",
            "optional": false,
            "customType": null as unknown as string
        },
        "datumBestand": {
            "type": "number",
            "optional": false,
            "customType": null as unknown as number,
            "serverName": "datum_bestand"
        },
        "peildatum": {
            "type": "number",
            "optional": false,
            "customType": null as unknown as number
        }
    },
    "primaryKey": ["jaar", "behandelendSpecialismeCd", "typerendeDiagnoseCd", "zorgproductCd", "peildatum"],
    "serverName": "fact_dbc"
} as const;
const factDbcProfielTable = {
    "name": "factDbcProfiel",
    "columns": {
        "jaar": {
            "type": "number",
            "optional": false,
            "customType": null as unknown as number
        },
        "behandelendSpecialismeCd": {
            "type": "string",
            "optional": false,
            "customType": null as unknown as string,
            "serverName": "behandelend_specialisme_cd"
        },
        "typerendeDiagnoseCd": {
            "type": "string",
            "optional": false,
            "customType": null as unknown as string,
            "serverName": "typerende_diagnose_cd"
        },
        "zorgproductCd": {
            "type": "string",
            "optional": false,
            "customType": null as unknown as string,
            "serverName": "zorgproduct_cd"
        },
        "zorgactiviteitCd": {
            "type": "string",
            "optional": false,
            "customType": null as unknown as string,
            "serverName": "zorgactiviteit_cd"
        },
        "zorgprofielklasseCd": {
            "type": "string",
            "optional": false,
            "customType": null as unknown as string,
            "serverName": "zorgprofielklasse_cd"
        },
        "aantalPat": {
            "type": "number",
            "optional": false,
            "customType": null as unknown as number,
            "serverName": "aantal_pat"
        },
        "aantalSubtraject": {
            "type": "number",
            "optional": false,
            "customType": null as unknown as number,
            "serverName": "aantal_subtraject"
        },
        "aantalZat": {
            "type": "number",
            "optional": false,
            "customType": null as unknown as number,
            "serverName": "aantal_zat"
        },
        "somAantalZat": {
            "type": "number",
            "optional": false,
            "customType": null as unknown as number,
            "serverName": "som_aantal_zat"
        },
        "versie": {
            "type": "string",
            "optional": false,
            "customType": null as unknown as string
        },
        "datumBestand": {
            "type": "number",
            "optional": false,
            "customType": null as unknown as number,
            "serverName": "datum_bestand"
        },
        "peildatum": {
            "type": "number",
            "optional": false,
            "customType": null as unknown as number
        }
    },
    "primaryKey": ["jaar", "behandelendSpecialismeCd", "typerendeDiagnoseCd", "zorgproductCd", "zorgactiviteitCd", "zorgprofielklasseCd", "peildatum"],
    "serverName": "fact_dbc_profiel"
} as const;
const refDiagnoseTable = {
    "name": "refDiagnose",
    "columns": {
        "diagnoseCd": {
            "type": "string",
            "optional": false,
            "customType": null as unknown as string,
            "serverName": "diagnose_cd"
        },
        "specialismeCd": {
            "type": "string",
            "optional": false,
            "customType": null as unknown as string,
            "serverName": "specialisme_cd"
        },
        "diagnoseOmschrijving": {
            "type": "string",
            "optional": false,
            "customType": null as unknown as string,
            "serverName": "diagnose_omschrijving"
        },
        "versie": {
            "type": "string",
            "optional": false,
            "customType": null as unknown as string
        },
        "datumBestand": {
            "type": "number",
            "optional": false,
            "customType": null as unknown as number,
            "serverName": "datum_bestand"
        },
        "peildatum": {
            "type": "number",
            "optional": false,
            "customType": null as unknown as number
        }
    },
    "primaryKey": ["diagnoseCd", "specialismeCd", "peildatum"],
    "serverName": "ref_diagnose"
} as const;
const refSpecialismeTable = {
    "name": "refSpecialisme",
    "columns": {
        "specialismeCd": {
            "type": "string",
            "optional": false,
            "customType": null as unknown as string,
            "serverName": "specialisme_cd"
        },
        "omschrijving": {
            "type": "string",
            "optional": false,
            "customType": null as unknown as string
        },
        "versie": {
            "type": "string",
            "optional": false,
            "customType": null as unknown as string
        },
        "datumBestand": {
            "type": "number",
            "optional": false,
            "customType": null as unknown as number,
            "serverName": "datum_bestand"
        },
        "peildatum": {
            "type": "number",
            "optional": false,
            "customType": null as unknown as number
        }
    },
    "primaryKey": ["specialismeCd"],
    "serverName": "ref_specialisme"
} as const;
const refZorgactiviteitTable = {
    "name": "refZorgactiviteit",
    "columns": {
        "zorgactiviteitCd": {
            "type": "string",
            "optional": false,
            "customType": null as unknown as string,
            "serverName": "zorgactiviteit_cd"
        },
        "omschrijving": {
            "type": "string",
            "optional": false,
            "customType": null as unknown as string
        },
        "zorgprofielklasseCd": {
            "type": "string",
            "optional": false,
            "customType": null as unknown as string,
            "serverName": "zorgprofielklasse_cd"
        },
        "versie": {
            "type": "string",
            "optional": false,
            "customType": null as unknown as string
        },
        "datumBestand": {
            "type": "number",
            "optional": false,
            "customType": null as unknown as number,
            "serverName": "datum_bestand"
        },
        "peildatum": {
            "type": "number",
            "optional": false,
            "customType": null as unknown as number
        }
    },
    "primaryKey": ["zorgactiviteitCd"],
    "serverName": "ref_zorgactiviteit"
} as const;
const refZorgproductTable = {
    "name": "refZorgproduct",
    "columns": {
        "zorgproductCd": {
            "type": "string",
            "optional": false,
            "customType": null as unknown as string,
            "serverName": "zorgproduct_cd"
        },
        "latijnOms": {
            "type": "string",
            "optional": false,
            "customType": null as unknown as string,
            "serverName": "latijn_oms"
        },
        "consumentOms": {
            "type": "string",
            "optional": false,
            "customType": null as unknown as string,
            "serverName": "consument_oms"
        },
        "declaratieVerzekerdCd": {
            "type": "string",
            "optional": false,
            "customType": null as unknown as string,
            "serverName": "declaratie_verzekerd_cd"
        },
        "declaratieOnverzekerdCd": {
            "type": "string",
            "optional": true,
            "customType": null as unknown as string,
            "serverName": "declaratie_onverzekerd_cd"
        },
        "versie": {
            "type": "string",
            "optional": false,
            "customType": null as unknown as string
        },
        "datumBestand": {
            "type": "number",
            "optional": false,
            "customType": null as unknown as number,
            "serverName": "datum_bestand"
        },
        "peildatum": {
            "type": "number",
            "optional": false,
            "customType": null as unknown as number
        }
    },
    "primaryKey": ["zorgproductCd"],
    "serverName": "ref_zorgproduct"
} as const;
// ...
export const schema = {
    "tables": {
        "factDbc": factDbcTable,
        "factDbcProfiel": factDbcProfielTable,
        "refDiagnose": refDiagnoseTable,
        "refSpecialisme": refSpecialismeTable,
        "refZorgactiviteit": refZorgactiviteitTable,
        "refZorgproduct": refZorgproductTable,
        "refZorgprofielklasse": refZorgprofielklasseTable
    },
    "relationships": {},
    "enableLegacyQueries": false,
    "enableLegacyMutators": false
} as const;

/**
 * Represents the Zero schema type.
 * This type is auto-generated from your Drizzle schema definition.
 */
export type Schema = typeof schema;
/**
 * Represents a row from the "factDbc" table.
 * This type is auto-generated from your Drizzle schema definition.
 */
export type FactDbc = Row<(typeof schema)["tables"]["factDbc"]>;